+++ changelog.md
@@ -1,3 +1,11 @@
+## 0.3.0
+- Configurable Settings! *Now requires Forge Config API Port*
+- Toggle appending the usability hint (Shift to expand tooltip line)
+- Toggle appending Block Interaction Range to relevant items
+- Toggle determining whether to collapse Enchantment Descriptions on gear by default
+- Allow configurable color of enchantment names/descriptions
+- Refactored logic for Block Interaction range tooltip (Thanks Daedelus!)
+
 ## 0.2.0
 - Fix Better Combat two-handed weapons not having the "Two-Handed" tooltip line
 - Color enchantment names based on their relation to the maximum enchantment value